Document

Description

Generic document-editing components.

Documentation

data EditingDocument c

Generic document editor with a dynamic viewport.q

editDocument

Arguments

:: FixedFontParser a c b

=> a
Parser used to break paragraphs into lines.

-> [UnparsedPara c]
List of unparsed paragraphs to be edited.

-> EditingDocument c
Document editor.

Create an editor for a document.

data EditAction c

Actions that modify data.

Constructors

InsertText [c]
Insert a block of characters.

DeleteText
Delete a single character.

data EditDirection

Modification direction, relative to the cursor.

Constructors

EditBefore
Apply the edit before the cursor.

EditAfter
Apply the edit after the cursor.

type EditorAction c = forall a. FixedFontEditor a c => a -> a

Any action that updates a FixedFontEditor.

class FixedFontEditor a c | a -> c where

Generic text editor for fixed-width fonts.

Minimal complete definition

editText, breakPara, moveCursor, getCursor, exportData

Methods

editText :: a -> EditAction c -> EditDirection -> a

Apply an EditAction.

breakPara :: a -> EditDirection -> a

Break the current paragraph at the cursor.

moveCursor :: a -> MoveDirection -> a

Apply a cursor movement.

getCursor :: a -> (Int, Int)

Get the cursor position, relative to the viewport if applicable.

exportData :: a -> [UnparsedPara c]

Export the modified data.

class DefaultBreak b => FixedFontParser a c b | a -> c b

Line parser for fixed-width fonts.

Minimal complete definition

setLineWidth, breakLines, renderLine

class FixedFontViewer a c | a -> c where

Generic editor viewport for fixed-width fonts.

Minimal complete definition

setViewSize, getViewSize, getVisible

Methods

setViewSize :: a -> (Int, Int) -> a

Sets the (width,height) size of the viewport. A width < 0 must disable line wrapping, and a height < 0 must disable vertical bounding.

getViewSize :: a -> (Int, Int)

Gets the (width,height) size of the viewport.

getVisible :: a -> [[c]]

Gets the visible lines in the viewport. This does not need to completely fill the viewport area, but it must not exceed it.

data MoveDirection

Actions that change the cursor position without changing data.

Constructors

MoveUp
Move up one line.

MoveDown
Move down one line.

MovePrev
Move backward one character.

MoveNext
Move forward one character.

MoveHome
Implementation-defined home operation.

MoveEnd
Implementation-defined end operation.

MovePageUp
Implementation-defined page-up operation.

MovePageDown
Implementation-defined page-down operation.

data UnparsedPara c

Single paragraph that has not been parsed into lines.

Constructors

UnparsedPara

Fields

upText :: [c]

The complete and uparsed paragraph data.

type ViewerAction c = forall a. FixedFontViewer a c => a -> a

Any action that updates a FixedFontViewer.

editorAppendAction :: [c] -> EditorAction c

Action for normal character insertion.

editorBackspaceAction :: EditorAction c

Action for the Backspace key.

editorDeleteAction :: EditorAction c

Action for the Delete key.

editorDownAction :: EditorAction c

Action for the down-arrow key.

editorEndAction :: EditorAction c

Action for the End key.

editorEnterAction :: EditorAction c

Action for the Enter key.

editorHomeAction :: EditorAction c

Action for the Home key.

editorInsertAction :: [c] -> EditorAction c

Action to insert after the cursor.

editorLeftAction :: EditorAction c

Action for the left-arrow key.

editorPageDownAction :: EditorAction c

Action for the PageDown key.

editorPageUpAction :: EditorAction c

Action for the PageUp key.

editorRightAction :: EditorAction c

Action for the right-arrow key.

editorUpAction :: EditorAction c

Action for the up-arrow key.

viewerResizeAction :: (Int, Int) -> ViewerAction c

Action to resize the viewport.