LineWrap

Description

Line-wrapping implementations. (See FixedFontParser for custom wrapping.)

Documentation

data BreakWords c

Wrapping policy that breaks lines based on words. Use breakWords to construct a new value.

data LineBreak

Line break type for a single paragraph line.

class WordSplitter a c | a -> c where

Word-splitting operations for use with BreakWords.

Methods

splitWord :: a -> Int -> Int -> [c] -> Maybe [Int]

Determine where to break a word.

The splitter can refuse to process the word by returning Nothing.

The segment sizes must provide space for a hyphen if appendHyphen extends the line.

Once the word has been split up by BreakWords, the segments are processed as follows:

The last segment is prepended to the next line to be parsed. This means that if the word is not split, it gets deferred to the next line.

If there are more segments, the first is appended to the current line being parsed.

All remaining segments are put on separate lines between the current and next lines.

isWordChar :: a -> c -> Bool

Predicate for characters that should be treated as a part of a word.

isWhitespace :: a -> c -> Bool

Predicate for detecting whitespace between words.

appendHyphen :: a -> [c] -> [c]

Append the canonical hyphen character to show word breaks.

breakExact :: BreakWords c

Wrapping policy that breaks at exactly the viewer width.

breakWords :: (Show a, WordSplitter a c) => a -> BreakWords c

Wrapping policy that breaks lines based on words.

lazyHyphen :: (WordChar c, HyphenChar c) => LazyHyphen c

Hyphenates words using simple aesthetics, without dictionary awareness.

lineBreakEnd :: LineBreak

The line is at the end of the paragraph.

lineBreakHyphen :: LineBreak

The line ends with a hyphenated word.

lineBreakSimple :: LineBreak

The line is nothing special.

noHyphen :: WordChar c => NoHyphen c

Avoids splitting words unless they are longer than a single line.