WEditor.Base.Parser

Description

Generic line-parsing functionality.

Documentation

class DefaultBreak b => FixedFontParser a c b | a -> c b where

Line parser for fixed-width fonts.

Minimal complete definition

setLineWidth, breakLines, renderLine

Methods

setLineWidth :: a -> Int -> a

Change the max line width used for parsing. A width of zero must result in breakLines skipping line breaks.

breakLines :: a -> [c] -> [VisibleLine c b]

Break the sequence into lines.

The following must hold for all possible inputs to a FixedFontParser p:

concat (map vlText (breakLines p line)) == line

Implement renderLine and tweakCursor to make visual adjustments (such as adding hyphens or indentation) if necessary.

renderLine :: a -> VisibleLine c b -> [c]

Render the line for viewing. Implement tweakCursor if renderLine changes the positions of any characters on the line.

tweakCursor :: a -> VisibleLine c b -> Int -> Int

Adjust the horizontal cursor position.