LineWrap

Description

Line-wrapping implementations. (See FixedFontParser for custom wrapping.)

Documentation

data BreakExact c

Wrapping policy that breaks at exactly the viewer width.

data BreakWords c

Wrapping policy that breaks lines based on words.

data LineBreak

Line break type for a single paragraph line.

type WordSplitter c

Arguments

= Int
Space available on the first line.

-> Int
Space available on new lines.

-> [c]
The word to break.

-> [[c]]
Word splits.

A function to split words, for use with BreakWords.

The word breaks must provide space for an additional hyphen character to be rendered if the word continues to the next line, i.e., all but the last item in the returned list should be one character shorter than the available space.

The splitter can refuse to process the word by returning an empty list.

Returning the whole word without breaks will bump the word to the next line. This must be avoided if the first line is full width, i.e., if the first two arguments are the same.

breakExact :: BreakExact c

Wrapping policy that breaks at exactly the viewer width.

breakWords :: (WordChar c, HyphenChar c) => WordSplitter c -> BreakWords c

Wrapping policy that breaks lines based on words.

lazyHyphen :: WordSplitter c

Hyphenates using simple aesthetics, without dictionary awareness.

lineBreakEnd :: LineBreak

The line is at the end of the paragraph.

lineBreakHyphen :: LineBreak

The line ends with a hyphenated word.

lineBreakSimple :: LineBreak

The line is nothing special.

noHyphen :: WordSplitter c

Avoids splitting words unless they are longer than a single line.