WEditor.Base.Parser

Description

Generic line-parsing functionality.

Documentation

class FixedFontParser a c | a -> c where

Line parser for fixed-width fonts.

Minimal complete definition

setLineWidth, breakLines, emptyLine, renderLine, splitLine

Associated Types

type BreakType a :: *

Type used to differentiate between line-break types.

Methods

setLineWidth :: a -> Int -> a

Change the max line width used for parsing. A width of zero must result in breakLines skipping line breaks.

breakLines :: a -> [c] -> [VisibleLine c (BreakType a)]

Break the sequence into lines.

The following must hold for all possible inputs to a FixedFontParser p:

concat (map vlText (breakLines p line)) == line

Implement renderLine and tweakCursor to make visual adjustments (such as adding hyphens or indentation) if necessary.

emptyLine :: a -> VisibleLine c (BreakType a)

A place-holder line for empty paragraphs.

renderLine :: a -> VisibleLine c (BreakType a) -> [c]

Render the line for viewing. Implement tweakCursor if renderLine changes the positions of any characters on the line.

tweakCursor :: a -> VisibleLine c (BreakType a) -> Int -> Int

Adjust the horizontal cursor position.

splitLine

Arguments

:: a

-> Int
Index to split at.

-> VisibleLine c (BreakType a)
Line to split.

-> (VisibleLine c (BreakType a), VisibleLine c (BreakType a))
New lines at (bottom,top) of previous/next paragraphs.

Split the line to create a paragraph break.

The following must hold for all possible inputs to a FixedFontParser p:

let (b,t) = splitLine p l in vlText l == vlText b ++ vlText t